Separated C++ wrappers into separate files in a backward-compatible way.
This makes both the C (.h) and C++ (.hpp) files read nicer.
They mutually #include each other, so no existing users should
need to change their code.
