+++ BUILD
@@ -81,6 +81,7 @@ cc_library(
         "upb/decode.h",
         "upb/encode.h",
         "upb/upb.h",
+        "upb/upb.hpp",
     ],
     copts = select({
         ":windows": [],
@@ -129,6 +130,7 @@ cc_library(
     ],
     hdrs = [
         "upb/def.h",
+        "upb/def.hpp",
         "upb/reflection.h",
     ],
     copts = select({
